+++ command/rm.go
@@ -53,7 +53,6 @@ var deleteCommand = &cli.Command{
 			c.Context,
 			c.Command.Name,
 			givenCommand(c),
-			s3opts(c),
 			c.Args().Slice()...,
 		)
 	},
@@ -64,7 +63,6 @@ func Delete(
 	ctx context.Context,
 	op string,
 	fullCommand string,
-	s3opts storage.S3Options,
 	src ...string,
 ) error {
 	srcurls, err := newURLs(src...)
@@ -73,7 +71,7 @@ func Delete(
 	}
 	srcurl := srcurls[0]
 
-	client, err := storage.NewClient(srcurl, s3opts)
+	client, err := storage.NewClient(srcurl)
 	if err != nil {
 		return err
 	}