+++ storage/storage.go
@@ -57,9 +57,9 @@ type Storage interface {
 
 // NewClient returns new Storage client from given url. Storage implementation
 // is inferred from the url.
-func NewClient(url *url.URL, opts S3Options) (Storage, error) {
+func NewClient(url *url.URL) (Storage, error) {
 	if url.IsRemote() {
-		return NewS3Storage(opts)
+		return NewS3Storage(url)
 	}
 
 	return NewFilesystem(), nil